requirements.txt
# AI Document Agent Dependencies

# Core AI/LLM Libraries
langchain>=0.1.0
openai>=1.0.0
anthropic>=0.8.0

# Document Processing
python-docx>=0.8.11
PyPDF2>=3.0.1
pdfplumber>=0.9.0

# WhatsApp/Communication (Future)
twilio>=8.0.0
flask>=2.3.0

# Environment and Configuration
python-dotenv>=1.0.0

# Utilities
pathlib2>=2.3.7
typing-extensions>=4.0.0

# Testing
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0